2023 Day 16
Lots of type state modelling today, but not too hard when you strip the
boilerplate away.

The tricky thing was to detect loops of beams, and terminate them before
they created endless loops. I did this by maintaining a shadow array
tracking what beams in what directions we had seen at each location. It
was important to distinguish between directions, as a beam going over
the same point in a different direction will be following a different
path. I did this by using the first 4 bits of a u8 at each location.

For part 2, I strongly suspect we can make it much faster still by using
a dynamic programming approach and caching the ensuing path from each
mirror / direction combination as we see it. The difficulty I had with
that is that my endless loop detection and termination might be
detecting an overlap with a bit of the path that predated the mirror
and so the found path won't be generically applicable to future visits
to that mirror from a different start location. An exercise for the
future dear reader