2023 Day 17
This was a pathfinding day with the extra spice that the movement is
constrained to only allow so many moves in one direction.

I realised almost straight away that I just needed Dijkstra or A* with
the visited set being a tuple of location, direction faced & how many
steps had been taken in a straight line. This is the heart of the
algorithm.

The problem took me almost a week to solve despite my initial insight as
I had assumed the starting direction was always East, but this was not a
valid assumption. Frustrated is an understatement right now...