2023 Day 18
This problem took me the longest time! Reading around after solving, it
looks like there's a simple algorithm called the shoelace algorithm that
can solve problems like this one relatively trivially. However my self
imposed rule of not searching for any hints prevented me from using
this. Also see below for some thoughts I've had about a simpler and
quicker way to improve the code I'm committing now.

The essence of this solution was to note that the area of the shape will
just be a series of rectangles. I stored the horizontal spans on an
initial traversal around the perimeter, plus what "height" they were.

I sorted the spans so they were in reverse height order. The sorting is
important as we want to find the matching pair of spans that are
adjacent in vertical space. The reverse order was so that I could pop
off the list of horizontals to look at the highest / top span.

Of course, spans won't neatly line up in pairs of fully matching in
horizontal space. I needed the concept of intersections, so I could
track how much a pair of intervals overlapped by and what their
residuals were after the overlap had been considered.

My method also double counted the vertical boundaries of neighbouring
rectangles, since the areas were inclusive. It was a bit of a faff, but
I also kept a vector of the vertical spans I had found so that we could
remove any double counts that were subsequently found

After being done with this it occurs to me that the area will just be
sum of the integrals of each horizontal span, which means I would be
able to get away with a single pass, plus each span would not be
dependent on any others, only itself. This ought to be significantly
faster and less complex to boot. I just couldn't be bothered to rewrite
for this fundamental change in approach