2023 Day 19
A fairly involved day, at least for my solution.

I noticed right away that it would probably be easier to build up a
collection of accepted parts, storing them as ranges of values for each
of the four categories.

Each test for each workflow would potentially partition an incoming
range in two (or either always be true / false, which happens when the
incoming range is already more restricted than the test being applied)

All we need do was recursively run through all the workflows starting
with an open range on all of the four conditions and stop when we either
get to Accept or Reject, with the former being saved as one of the
accept ranges

I also indexed the workflows to speed up the recursive search for the
next workflow to apply