2023 Day 20
I went to town a little on this one and separated it all out into neat
modules with their own file.

For part 1, I converted the state of the modules into a u64 that stored
the state of the flip flops and the inputs to the conjunctions as
different bits in the number. This was only needed for the test cases as
the main problem actually had a cycle WAY higher than 1,000. Which
brings us to part 2...

For part 2 I cheated really as I looked for the first instance of each
of the output conjunction modules flipping high and then assumed that
this is finding the cycle of when they flip, which it was, but this is
not guaranteed. Also I should have really taken the LCM of all of the
input cycles but I manually verified that I had been given prime numbers
for each of the cycles and I think it a fair bet that all inputs for
this day will follow this pattern