2023 Day 21
Today's solution is a little hacky and relies on being given a nicely
formatted problem. I think it's a fair bet we get this but I can't
guarantee it.

The key insight here is that the spaces that an elf can reach look like
a diamond centred on the middle start point, as long as the elf travels
65 steps + some multiple of the map width (131), then garden squares he
can reach are conveniently exactly covered by an integer number of
smaller diamond patterns of garden & rock.

The squares he can reach oscillate between odd and even step counts
since he can only ever get to the squares one away from the start point
with an odd number of steps, and the same point applies an even number
of steps to the squares two away. This pattern repeats outward
indefinitely.

The smaller garden patterns the elf can reach, are made of two types:
the inner diamond that is visible if the input data is just printed &
the outer diamond that is made by pasting together the 4 corners of the
input pattern. Each of these two patterns has an even and an odd step
count version of which squares can be reached when fully covered.

When working out how many squares are covered by a certain number of
steps it suffices to fully step out to the pattern repeated 9 times in a
3 x 3 grid. This will cover 4 lots of the inner odd pattern, 1 lot of
the inner even pattern & 2 lots each of the odd and even outer patterns;
      io
   oe    oo
io    ie    io
   oo    oe
      io

To work out the steps covered, we don't need to count squares we have
already been to since an oscillating pattern of stepping from one square
to it's neighbour and back again will always find the source square
again. We therefore maintain a frontier of new squares not yet visited
when searching out over the initial 3 x 3 grid. We have to maintain two
sets of visited squares due to the oscillating nature of the step counts
already discussed.

All this is calculated by the `run` function. Part 1 is then just the
inner even count. For part 2, we can see that if the pattern is repeated
`n` times out in any direction that we will have `n^2` copies of the
inner even pattern, `(n + 1)^2` copies of the inner odd pattern & `n *
(n + 1)` copies of the outer even and outer odd patterns combined. So
part 2 is reduced to applying that formula to the three base numbers
found over the 3 x 3 grid