2023 Day 22
This day reduced down to a directed acyclic graph of which bricks were
above and below each brick. Rust makes referential graphs hard work, so
I built arrays of usize indexes, which could be used to index into the
arrays to find the next nodes.

To set things up, I sorted all the input data of bricks into ascending z
value. Since bricks cannot fall through each other this one sorting
guaranteed that after falling all the bricks, all the bricks below would
always be below in the array and likewise bricks above would be after in
the array. This is true even though after falling the bricks would no
longer be sorted in z order

To fall the stack, I kept an x/y plane memory of the current surface of
the fallen stack. Each next brick then falls as far as the surface
directly below the brick allows. In falling the brick we update it's z
value to the new lower amount and then update the surface for the top of
the newly fallen brick

I used a second pass through the array of fallen bricks to find which
bricks are directly touching above and below the given brick. Noticing
that each brick is at most 5 units long the Adjacent array has 5
potential slots for bricks that are touching above and below. As
discussed above, rather than using direct references, I used arrays of
indexes to the given bricks since that wouldn't create any borrowing
headaches

Having built the graph, we can actually ditch the position of the bricks
themselves since all that matters is which brick is next to which.

For Part 2, we can build up a hashset of the affected bricks above each
brick by cycling through all the bricks above and adding any bricks
whose antecedents are all in the set of affected bricks. I replaced the
Hashset with my own fixed array. Given we're working in the space 0..N
using a fixed array will just be much quicker than using heap storage
and a hash function for each insert and lookup