2023 Day 23
This was a bit of a journey for me. In the end I realised that the
supplied ASCII maze could actually be reduced to a graph of 34 nodes.
Further those nodes were connected to one another in quite a specific
pattern:

 0
 |
 1 -  3 -  6 - 10 - 15 - \
 |    |    |    |    |    |
 2 -  5 -  9 - 14 - 20 - 21
 |    |    |    |    |    |
 4 -  8 - 13 - 19 - 25 - 26
 |    |    |    |    |    |
 7 - 12 - 18 - 24 - 29 - 30
 |    |    |    |    |    |
11 - 17 - 23 - 28 - 32 - 33
 |    |    |    |    |    |
 \ - 16 - 22 - 27 - 31 - 34
                          |
                          END

Finally, for part 1, the downhill arrows only & always funnelled the
direction from a node South and East, so that part 1 was just a much
simplified walk across the grid.

The initial phase therefore was to walk the actual maze, find the
junction nodes and calculate the distances between the nodes. After this
first pass we could ignore the supplied input and just work with the
graph of nodes and distances.

Since we only had 34 nodes, we could use the bits of a u64 to
efficiently track which nodes had been visited on a given walk.

The max path of a graph in general requires every possible path to be
searched but in this case we could exclude turning backwards on any edge
node, since it can be seen we would cut our path off from the end by
making such a turn. This significantly reduces the number of paths to be
searched.

Finally, the searching of the paths are independent, so we can thread
the search for another performance boost. I only ran with 4 threads in
the end but it would be quite simple to step up to 10 threads (the ways
that the first 4 steps could be taken).