2024 Day 1
Simple enough. I had a prep stage that sorted the inputs into a pair of sorted
number lists, since that was useful for both parts.

For part 2, I went back to C style! As we could iterate through both lists in
parallel to get the answer in a single pass of both lists.