2024 Day 2
I found this surprisingly involved for part 2. In the end I decided we needed
the first 4 numbers off the rank to determine a consensus on direction, allowing
for any 1 number to be wrong and ignorable: if the first 3 numbers go up, then
down; how can we decide which of the first three numbers to ignore without going
to the fourth to settle it?

I then made life simpler for myself by writing a function that would sort out if
any 3 numbers were strictly in order. We could then use that to iterate through
the 4 combinations of missing one number out from the first four to see which
combinataion (if any) had three numbers in consensus and we could then use that
to iterate through the rest of the line.