2024 Day 3
Today was crying out for nom, but since I'm trying to be vanilla I coded my own.
I went with a state machine to tell where we were up to in the parsing.

It's coded up as two separate passes through the text currently, which I may
choose to come back and refactor. If you're reading this though it looks like I
didn't!