2024 Day 4
I stored the input data as a 1D byte array, which I made look like a 2D grid by
indexing into it using a row and column. After that it was a case of creating a
Point struct and a Direction enum, that I poshed up by letting you add one to
another and also allowed for multiplication of the direction. This was so we
could "move" around the grid in various directions.

The other trick I enjoyed in this attempt was having functions that returned
Iterators, which made combining them all together into one simple functional
pass to calculate the answer very easy.