2024 Day 5
I ended up approaching this by noticing that the rules data involved 49 distinct
pages (from the span of two digit numbers 10 to 99). Every page was in front of
24 other pages and behind 24 pages, i.e. there was a sort order for every page,
which said if it was ahead or behind every other page. This is equivalent to
saying that every rule involving a page, is set given that that page is in the
middle of all pages.

Of course not all pages can be in the middle of a unique sort order, in fact
clearly only one can! So what this means is that the full set of rules are
mutually exclusive, but that they define a ring sort order.

So what I did was to first process the rules to find this ring order. I did this
by picking a random page to go first (any page could be first in a ring order).
Having picked this first page, the next page will be the page with a list of
pages that come after it that is only one different to the initial page. In this
way we can iterate through the pages finding the next page in the ordered list.

In fact we can go a bit faster than that by observing that the new page in the
list of 24 'after' pages for the following page must be the page that is last in
that list of 24 pages, or the first page in the second half of the list for the
initial page. With this info we only need find half of the following pages,
building up the second half order from the new pages in the after lists on the
first half pages that are found.

With a concrete order, to test if a sample page list was in the correct order,
we iterated through the sample pages and 24 of the concrete order (starting with
the initial sample page) and if we find all the sample pages in order then it is
ordered, otherwise it is not.

Re ordering was a bit more of a faff but we found the list of orders of the
sample pages, sorted them, picked the middle one and then converted that sort
order number back into a page number.

Simples!