2025 Day 2
The "obvious" solution today would be to iterate over every number in
the given ranges, testing if it is an invalid number and then add those
resultant numbers. This bothered me as an approach as we only need a
fraction of those numbers.

By looking at the top n digits (in base 10) of the start and end of the
range we can determine the invalid numbers by just iterating between
these "base" numbers and then compounding them up into invalid numbers.

For part 1, this just involved doubling the numbers. For part 2, we need
to iterate over all the possible repeats. In part 2 we also need to
guard against adding the same number over and over, since 222222 is
double 222 and also triple 22, and sextuple 2!